Fix update script

+++ update.sh
@@ -5,7 +5,7 @@ dir=`pwd`
 if [ $USER != "cmsdqm" ]; then
     echo "Script must be run as user: cmsdqm"
     echo "Switch user: sudo su cmsdqm"
-    exit -1
+    exit 1
 fi
 
 # Creates a new directory to clone the source files into.
@@ -15,15 +15,13 @@ git clone https://github.com/cms-dqm/CentralHDQM $src
 cd $src
 git checkout master
 
-# Copy the .env file
+# Copy the .env file: It's expected to be found in /data/hdqm.
+if [! -f ../.env]; then
+    echo "A file named .env is expected to be in the same directory with update.sh; Create it and rerun the script"
+    exit 1
+fi
 cp ../.env ./backend/
 
-# TODO: Why is CMSSW needed?
-export RELEASE=/cvmfs/cms.cern.ch/slc7_amd64_gcc10/cms/cmssw/CMSSW_12_4_5
-source /cvmfs/cms.cern.ch/cmsset_default.sh
-cd $RELEASE
-eval `scramv1 runtime -sh`
-cd -  # Go back to $src
 
 # Create a venv, activate it, install requirements.
 PYTHON=`(which python3)`
@@ -31,11 +29,14 @@ $PYTHON -m venv venv
 source venv/bin/activate
 $PYTHON -m pip install -r requirements.txt -U --no-cache-dir
 
+# Add ROOT to PATH (includes PyROOT)
+. /cvmfs/sft.cern.ch/lcg/app/releases/ROOT/6.24.08/x86_64-centos7-gcc48-opt/bin/thisroot.sh
+
 # Get back to main dir to switch the link
 cd $dir
 
 # Softlink latest source files to a link named "current"
 ln -s -f -n $src current
 
-echo "New version started! Don't forget to restart: sudo systemctl restart hdqm2.service"
+echo "New version started! Don't forget to restart: sudo systemctl restart hdqm.service"
